requirements.txt
Django==1.9.2
django-cors-headers==1.1.0
djangorestframework==3.3.2
nose==1.3.7
psycopg2==2.6.1
wheel==0.29.0
